+++ src/main/java/jp/kusumotolab/kgenprog/fl/Tarantula.java
@@ -8,7 +8,7 @@ public class Tarantula extends SpectrumBasedFaultLocalization {
 
   @Override
   protected double formula(final double ef, final double nf, final double ep, final double np) {
-    return (ef / (double) (ef + nf)) / (ef / (double) (ef + nf) + ep / (double) (ep + np));
+    return (ef / (ef + nf)) / (ef / (ef + nf) + ep / (ep + np));
   }
 
 }
